+++ crates/miden-lib/src/errors/note_script_errors.rs
@@ -10,6 +10,17 @@ use crate::errors::MasmError;
 // NOTE SCRIPT ERRORS
 // ================================================================================================
 
+/// Error Message: "failed to reclaim P2IDE note because the reclaiming account is not the sender"
+pub const ERR_P2IDE_RECLAIM_ACCT_IS_NOT_SENDER: MasmError = MasmError::from_static_str("failed to reclaim P2IDE note because the reclaiming account is not the sender");
+/// Error Message: "P2IDE reclaim is disabled"
+pub const ERR_P2IDE_RECLAIM_DISABLED: MasmError = MasmError::from_static_str("P2IDE reclaim is disabled");
+/// Error Message: "failed to reclaim P2IDE note because the reclaim block height is not reached yet"
+pub const ERR_P2IDE_RECLAIM_HEIGHT_NOT_REACHED: MasmError = MasmError::from_static_str("failed to reclaim P2IDE note because the reclaim block height is not reached yet");
+/// Error Message: "failed to consume P2IDE note because the note is still timelocked"
+pub const ERR_P2IDE_TIMELOCK_HEIGHT_NOT_REACHED: MasmError = MasmError::from_static_str("failed to consume P2IDE note because the note is still timelocked");
+/// Error Message: "P2IDE note expects exactly 4 note inputs"
+pub const ERR_P2IDE_WRONG_NUMBER_OF_INPUTS: MasmError = MasmError::from_static_str("P2IDE note expects exactly 4 note inputs");
+
 /// Error Message: "P2IDR's reclaimer is not the original sender"
 pub const ERR_P2IDR_RECLAIM_ACCT_IS_NOT_SENDER: MasmError = MasmError::from_static_str("P2IDR's reclaimer is not the original sender");
 /// Error Message: "P2IDR can not be reclaimed as the transaction's reference block is lower than the reclaim height"
@@ -19,8 +30,8 @@ pub const ERR_P2IDR_WRONG_NUMBER_OF_INPUTS: MasmError = MasmError::from_static_s
 
 /// Error Message: "P2ID's target account address and transaction address do not match"
 pub const ERR_P2ID_TARGET_ACCT_MISMATCH: MasmError = MasmError::from_static_str("P2ID's target account address and transaction address do not match");
-/// Error Message: "P2ID script expects exactly 2 note inputs"
-pub const ERR_P2ID_WRONG_NUMBER_OF_INPUTS: MasmError = MasmError::from_static_str("P2ID script expects exactly 2 note inputs");
+/// Error Message: "P2ID note expects exactly 2 note inputs"
+pub const ERR_P2ID_WRONG_NUMBER_OF_INPUTS: MasmError = MasmError::from_static_str("P2ID note expects exactly 2 note inputs");
 
 /// Error Message: "SWAP script requires exactly 1 note asset"
 pub const ERR_SWAP_WRONG_NUMBER_OF_ASSETS: MasmError = MasmError::from_static_str("SWAP script requires exactly 1 note asset");