+++ crates/miden-lib/src/note/utils.rs
@@ -1,7 +1,8 @@
 use miden_objects::{
-    NoteError, Word,
+    Felt, NoteError, Word,
     account::AccountId,
     asset::Asset,
+    block::BlockNumber,
     note::{NoteExecutionMode, NoteInputs, NoteRecipient, NoteTag, NoteType},
 };
 
@@ -17,6 +18,31 @@ pub fn build_p2id_recipient(
 ) -> Result<NoteRecipient, NoteError> {
     let note_script = WellKnownNote::P2ID.script();
     let note_inputs = NoteInputs::new(vec![target.suffix(), target.prefix().as_felt()])?;
+
+    Ok(NoteRecipient::new(serial_num, note_script, note_inputs))
+}
+
+/// Creates a [NoteRecipient] for the P2IDE note.
+///
+/// Notes created with this recipient will be P2IDE notes consumable by the specified target
+/// account.
+pub fn build_p2ide_recipient(
+    target: AccountId,
+    reclaim_block_height: Option<BlockNumber>,
+    timelock_block_height: Option<BlockNumber>,
+    serial_num: Word,
+) -> Result<NoteRecipient, NoteError> {
+    let note_script = WellKnownNote::P2IDE.script();
+
+    let reclaim_height_u32 = reclaim_block_height.map_or(0, |bn| bn.as_u32());
+    let timelock_height_u32 = timelock_block_height.map_or(0, |bn| bn.as_u32());
+
+    let note_inputs = NoteInputs::new(vec![
+        target.suffix(),
+        target.prefix().into(),
+        Felt::new(reclaim_height_u32 as u64),
+        Felt::new(timelock_height_u32 as u64),
+    ])?;
 
     Ok(NoteRecipient::new(serial_num, note_script, note_inputs))
 }